GS::GitStatistics -> GS::CLI
Due to the confusing nature of a module and a class
having the same name, rename the `GitStatistics` class
to `CLI`

The `Cli#execute` method was also doing way too much. Break
down the complexity into smaller pieces to reduce comment
noise and make eventual integration testing easier to
accomplish.